[ROUTING][COURSES] Adding custom filtering and did code refactor - Including pagination, filtering, sorting the course results - Did code refactor of the bootcamp's controller - Created new middleware with shared logic like: pagination, filtering and sorting the results
