+++ src/controllers/bootcamps.js
@@ -8,74 +8,10 @@ const asyncHandler = require('../utils/asyncHandler');
 
 // @see: https://docs.mongodb.com/manual/reference/operator/query/
 // @see: https://mongoosejs.com/docs/tutorials/query_casting.html
-// TODO Code-refactor because the code below potentially will be duplicate code of other resources
+
 exports.getAllBootCamps = asyncHandler(async (req, res, next) => {
-    let queryString = null;
-    const { query } = req;
-    const requestQuery = { ...query }
-    const removeFields = [
-        'selectBy', 'sortBy',
-        'page', 'limit'
-    ];
-
-    // Removing fields from the query params
-    for (let value of removeFields) delete requestQuery[value]
-
-    queryString = JSON.stringify(requestQuery);
-
-    // Parsing query params as a query language in MongoDB
-    // Example params:  { "query": { "averageCost": { "$gte": "1000" } } }
-    const queryParams = JSON.parse(
-        queryString.replace(/\b(gt|gte|lt|lte|in)\b/g, match => `$${match}`)
-    )
-
-    let querySelect = BootcampModule.find(queryParams).populate({
-        path: 'courses',
-        select: 'title -bootcamp'
-    })
-
-    // Selecting data by particular fields
-    if (query.hasOwnProperty('selectBy')) {
-        const fields = query.selectBy.split(',').join(' ')
-        querySelect = querySelect.select(fields)
-    }
-
-    // Sort data by particular fields
-    if (query.hasOwnProperty('sortBy')) {
-        const sortBy = query.sortBy.split(',').join(' ')
-        querySelect = querySelect.sort(sortBy)
-    } else {
-        querySelect = querySelect.sort('-createdAt')
-    }
-
-    const page = parseInt(query.page) || 1;
-    const limit = parseInt(query.limit) || 20;
-    const startIdx = (page - 1) * limit
-    const endIdx = page * limit;
-    const pagination = null;
-    const total = await BootcampModule.countDocuments();
-
-    if (endIdx < total) {
-        pagination.next = { page: page + 1}
-        pagination.limit = limit
-    }
-
-    if (startIdx > 0) {
-        pagination.prev = { page: page - 1}
-        pagination.limit = limit
-    }
-
-    querySelect = querySelect.skip(startIdx).limit(limit)
-
-    const bootcampArray = await querySelect;
     res.status(200);
-    res.json({
-        success: true,
-        pagination,
-        size: bootcampArray.length,
-        query: queryParams,
-        data: bootcampArray,
-    })
+    res.json(res.advanceFiltering)
 });
 
 // @desc:   Get single bootcamp
